pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
ipython = "^8.27.0"
djangorestframework = "^3.15.2"
redis = "^5.0.8"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-filter = "^24.3"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
